refactor(cli): extract query orchestration to reduce handler complexity
Create apps/cli/src/query.ts with orchestration functions:
- validateQueryOptions: Validate mutually exclusive options
- buildCliQueryContext: Parse options and resolve all filters
- buildCliQueryFilters: Build query options for queryEntries()
- applyClientSideFilters: Apply PR/author/mine/reviews filters
- renderCliQueryOutput: Render summary/json/human-readable output
- executeCliQuery: Main entry point coordinating all phases

Both list.ts and index.ts handlers now delegate to executeCliQuery(),
reducing complexity from ~65 to ~10.

Complexity analysis:
- list.ts: 65 -> ~10 (handler only calls executeCliQuery)
- index.ts: 63 -> ~10 (handler only calls executeCliQuery)
- MCP handleQuery: 41 (unchanged, acceptable warning)
